resident ok

+++ src/components/Content/Resident/index.js
@@ -9,16 +9,13 @@ import { Delete } from "./delete";
 import ApartmentService from "../../../services/ApartmentService";
 import ResidentService from "../../../services/ResidentService";
 
-const aptOptions = [
-  "101", "102", "103", "104", "201", "202", "203", "204"
-];
-
 const selectBlocks = ["Não há blocos"];
 
 const initialEditState = {
-  numeroApto: "",
-  blocoApto: "",
-  vagaGaragem: ""
+  nome: "",
+  email: "",
+  telefone: "",
+  dataNascimento: ""
 };
 
 export const Resident = ({ tabActive }) => {
@@ -27,17 +24,42 @@ export const Resident = ({ tabActive }) => {
   const [apartments, setApartments] = useState([]);
   const [numbersApt, setNumbersApt] = useState([]);
   const [createData, setCreateData] = useState({});
-  const [searchResult, setSearchResult] = useState("");
+  const [searchTerm, setSearchTerm] = useState("");
+  const [searchResult, setSearchResult] = useState([]);
   const [editData, setEditData] = useState(initialEditState);
 
   useEffect(() => {
     getApartments();
   }, []);
 
-  const handleSelectApartment = id => {
-    const selectApt = apartments.find(it => it.numeroApto === id);
+  useEffect(() => {
+    setApartment({});
+    getApartments();
+    setSearchResult([]);
+  }, [tabActive]);
 
-    getApartment(selectApt.id);
+  useEffect(() => {
+    const aptResident = apartment && apartment.morador !== undefined;
+    setSearchResult(aptResident ? [apartment.morador] : []);
+  }, [apartment]);
+
+  useEffect(() => {
+    const results = residents.filter(person => person.nome.includes(searchTerm));
+    setSearchResult(results);
+  }, [searchTerm]);
+
+  // HANDLERS
+  // _____________
+  const handleSelectApartment = id => {
+    if (id) {
+      const selectApt = apartments.find(it => it.numeroApto === id);
+  
+      getApartment(selectApt.id);
+    }
+  };
+  
+  const handleSearchResident = e => {
+    setSearchTerm(e.target.value)
   };
 
   // HTTP METHODS
@@ -91,8 +113,6 @@ export const Resident = ({ tabActive }) => {
       }
     };
 
-    console.log(body)
-
     ResidentService.createResident(body).then(response => {
       toast.success(response.data.mensagem, {
         position: toast.POSITION.TOP_CENTER
@@ -105,7 +125,43 @@ export const Resident = ({ tabActive }) => {
     ));
   };
 
+  const editResident = (e, id) => {
+    e.preventDefault();
+
+    if (!id) {
+      return toast.error("ID inválido! Selecione um morador e tente novamente.", {
+        position: toast.POSITION.TOP_CENTER
+      });
+    };
+
+    const body = {
+      "morador": {
+        ...editData,
+        "apartamento": {
+          "id": apartment.id
+        }
+      }
+    }
+
+    ResidentService.editResident(id, body).then(response => {
+      toast.success(response.data.mensagem, {
+        position: toast.POSITION.TOP_CENTER
+      });
+      return setResidents([ ...residents, editData ]);
+    }).catch(e => (
+      toast.error(e.mensagem, {
+        position: toast.POSITION.TOP_CENTER
+      })
+    ));
+  };
+
   const deleteResident = id => {
+    if (!id) {
+      return toast.error("ID inválido! Selecione um morador e tente novamente.", {
+        position: toast.POSITION.TOP_CENTER
+      });
+    };
+
     ResidentService.deleteResident(id).then(response => {
       toast.success(response.data.mensagem, {
         position: toast.POSITION.TOP_CENTER
@@ -122,14 +178,12 @@ export const Resident = ({ tabActive }) => {
     <div className="content">
       {tabActive === "search" && 
         <Search
-          apartment={apartment}
-          residents={residents}
           numbersApt={numbersApt}
-          apartments={apartments}
+          searchTerm={searchTerm}
           selectBlocks={selectBlocks}
           getResidents={getResidents}
           searchResult={searchResult}
-          setSearchResult={setSearchResult}
+          handleSearchResident={handleSearchResident}
           handleSelectApartment={handleSelectApartment}
         />
       }
@@ -146,15 +200,28 @@ export const Resident = ({ tabActive }) => {
       }
       {tabActive === "edit" && 
         <Edit
-          apartments={numbersApt}
+          editData={editData}
+          numbersApt={numbersApt}
+          setEditData={setEditData}
+          apartment={apartment}
+          editResident={editResident}
           selectBlocks={selectBlocks}
+          searchResult={searchResult}
+          initialEditState={initialEditState}
+          handleSelectApartment={handleSelectApartment}
         />
       }
       {tabActive === "delete" && 
         <Delete
+          residents={residents}
           numbersApt={numbersApt}
+          searchTerm={searchTerm}
           selectBlocks={selectBlocks}
+          getResidents={getResidents}
+          searchResult={searchResult}
           deleteResident={deleteResident}
+          handleSearchResident={handleSearchResident}
+          handleSelectApartment={handleSelectApartment}
         />
       }
     </div>
